+++ ne_version.h
@@ -1,3 +1,3 @@
 ﻿#pragma once
-#define NE_PUBLISHED L"27.05.2026 10:40"
-#define NE_VERSION   L"2026.05.27.10"
+#define NE_PUBLISHED L"27.05.2026 11:35"
+#define NE_VERSION   L"2026.05.27.11"